+++ src/LaravelBackupPanelServiceProvider.php
@@ -2,6 +2,7 @@
 
 namespace PavelMironchik\LaravelBackupPanel;
 
+use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Config;
 use Illuminate\Support\Facades\Route;
 use Illuminate\Support\ServiceProvider;
@@ -29,6 +30,10 @@ class LaravelBackupPanelServiceProvider extends ServiceProvider
             ], 'laravel-backup-panel-views');
 
             $this->publishes([
+                __DIR__.'/../resources/lang' => lang_path('vendor/laravel_backup_panel'),
+            ], 'laravel-backup-panel-translations');
+
+            $this->publishes([
                 __DIR__.'/../stubs/LaravelBackupPanelServiceProvider.php.stub' => app_path('Providers/LaravelBackupPanelServiceProvider.php'),
             ], 'laravel-backup-panel-provider');
 
@@ -39,12 +44,17 @@ class LaravelBackupPanelServiceProvider extends ServiceProvider
 
         Route::group([
             'prefix' => Config::string('laravel_backup_panel.path'),
-            'middleware' => ['web', Authenticate::class],
+            'middleware' => [
+                'web',
+                Authenticate::class,
+                ...Config::array('laravel_backup_panel.middleware'),
+            ],
         ], function () {
             $this->loadRoutesFrom(__DIR__.'/../routes/web.php');
         });
 
         $this->loadViewsFrom(__DIR__.'/../resources/views', 'laravel_backup_panel');
+        $this->loadTranslationsFrom(__DIR__.'/../resources/lang', 'laravel_backup_panel');
     }
 
     /**
@@ -54,5 +64,7 @@ class LaravelBackupPanelServiceProvider extends ServiceProvider
     public function register(): void
     {
         $this->mergeConfigFrom(__DIR__.'/../config/laravel_backup_panel.php', 'laravel_backup_panel');
+
+        LaravelBackupPanel::auth(static fn (Request $request): bool => false);
     }
 }