+++ tests/TestCase.php
@@ -2,8 +2,6 @@
 
 namespace PavelMironchik\LaravelBackupPanel\Tests;
 
-use Illuminate\Http\Request;
-use PavelMironchik\LaravelBackupPanel\LaravelBackupPanel;
 use PavelMironchik\LaravelBackupPanel\LaravelBackupPanelServiceProvider;
 use Spatie\Backup\BackupServiceProvider;
 
@@ -20,12 +18,5 @@ class TestCase extends \Orchestra\Testbench\TestCase
             LaravelBackupPanelServiceProvider::class,
             BackupServiceProvider::class,
         ];
-    }
-
-    protected function setUp(): void
-    {
-        parent::setUp();
-
-        LaravelBackupPanel::auth(static fn (Request $request): bool => false);
     }
 }